added data relationships between the users and posts and added a bit of validation to the resolvers to reifne the resolvers

+++ src/resolvers/mutations/Post.ts
@@ -8,15 +8,30 @@ interface PostArgs {
   id: string;
 }
 
+interface PostCreateInputArgs {
+  options: {
+    title: string;
+    text: string;
+  };
+}
+
 export const postResolvers = {
   //MUTATION TO CREATE POST
   postCreate: async (
     _: any,
-    { title }: PostArgs,
-    { prisma }: Context
+    { options }: PostCreateInputArgs,
+    { prisma, req }: Context
   ): Promise<Post | null> => {
+    if (!req.session.userId) {
+      throw new Error('unauthenticated user');
+    }
+    const { text, title } = options;
     const createdPost = await prisma.post.create({
-      data: { title: title },
+      data: {
+        title: title,
+        text: text,
+        creatorId: req.session.userId,
+      },
     });
     return createdPost;
   },
@@ -24,8 +39,11 @@ export const postResolvers = {
   postUpdate: async (
     _: any,
     { id, title }: PostArgs,
-    { prisma }: Context
+    { prisma, req }: Context
   ): Promise<Post | null> => {
+    if (!req.session.userId) {
+      throw new Error('unauthenticated user');
+    }
     const post = await prisma.post.findUnique({
       where: {
         id: Number(id),
@@ -49,8 +67,11 @@ export const postResolvers = {
   postDelete: async (
     _: any,
     { id }: PostArgs,
-    { prisma }: Context
+    { prisma, req }: Context
   ): Promise<boolean> => {
+    if (!req.session.userId) {
+      throw new Error('unauthenticated user');
+    }
     await prisma.post.delete({
       where: { id: Number(id) },
     });
